requirements.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
fastapi==0.110.1
h11==0.14.0
idna==3.6
openslide-python==1.3.1
pillow==10.3.0
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
python-magic==0.4.27
python-multipart==0.0.9
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
